### Release 1.0.7 (2011/04/19, 00:02:30)
-------------------------------------------
SUMMARY OF CHANGES:

Fixed hang in the File Revision History form when diffing

Added a separate git execution function that does not block.
It is now possible to have multiple files revision dialogs
and each with multiple opened diff tools to see various
revisions.

Fixing edit of a file in Linux

Since we use FullPath property of a treeview, set the
platform-specific path separator character.

Added extra time-slice in Run()

There are known problems with the async output not being flushed as the
thread exits. Releasing a time-slice using DoEvents seems to fix
the problem in this particular setting.

Added mutex around global Refresh

Protect Refresh chain with a simple exit mutex, so that the F5 key
(update) does not start a re-entrant refresh chain. Although the
main GUI app is single-threaded, some panels refresh functions
are calling GitRun() which in turn spawns external async process
during which time we can end up with multiple threads trying to
refresh.