Refresh view pane on the file edit
Setup a file watcher on the left pane and trigger a refresh
when a file being edited was saved/modified. It was necessary
to create a file watcher since the editor is called
asynchronously.
This applies both to the double-clicking of a file and to the
explicit file edit.
TODO: There are some instances in Windows where the file
change would cause a trigger.