### Release 1.0.15 (2012/02/09, 00:47:05)
-------------------------------------------
SUMMARY OF CHANGES:

Add the setting panes to edit gitconfig files

There are 2 files: a global gitconfig file (in the user
directory) and a config file per each repo. This change
adds the widget tabs in settings and repo edit to be
able to edit those files directly.

Add UserHome global string.

Set the new version check timeout to 4 sec

This prevents ocasional hang when the web site does
not reply on time.

In the tools menu, add buttons with the help displaying
and selecting tokens for argument and directory macros:

%r - Root directory of the current git repository
%u - User name
%e - User email
%b - Active branch of the current git repository
%f - Single selected file in the left pane
%F - List of selected files in the left pane
%d - Single selected directory in the left pane
%D - List of selected directories in the left pane