### Release 1.0.19 (2012/10/12, 21:46:24)
-------------------------------------------
SUMMARY OF CHANGES:

Modified release script

To use "git.exe" instead of "git.cmd". It used to work before, and
now it doesn't for some reason (?). The exe works.

Fixing a failing commit on some hosts

If a temp path contains spaces, commit would fail. Adding quotes around
the temp file name in the commit function to fix that case.