Automatically download and install KDiff3 diff utility
On startup, GitForce checks for the presence of a diff tool
and will offer to download and install it for the user.
This is only for Windows and it downloads KDiff3 tool.

Due to the difficulty in parsing the exact web path to the
exe file in SF.NET, I copied the most recent version to my
Github project and am getting it from there:

https://github.com/gdevic/KDiff3