### Release 1.0.22 (2012/11/17, 00:03:58)
-------------------------------------------
SUMMARY OF CHANGES:

Automatically download and install KDiff3 diff utility

On startup, GitForce checks for the presence of a diff tool
and will offer to download and install it for the user.
This is only for Windows and it downloads KDiff3 tool.

Due to the difficulty in parsing the exact web path to the
exe file in SF.NET, I copied the most recent version to my
Github project and am getting it from there:

https://github.com/gdevic/KDiff3

Automatically download and install msysgit

On Windows, msysgit is the defualt git tool. On startup, GitForce
checks for its presence and will offer to download and install it
for the user. This code does that.

Improvements to checking for a new version

Print the results of a check on the status pane
Moved check to later in the init process

Fixing the refresh chain

Some of the Refresh buttons were not connected completely in the sense
that they did not refresh what would the user expect. F5 did it
correctly but at the global scale.