Optionally show tabs on the right panel.
This change is based on the original change by splintor:
"Show right panels in a tab control": 'Showing the right panels in a tab
control makes it more visible that it can show'

His change has been slightly modified to make the tabs optional.

Showing tabs is optional (by default it is enabled). There is a new
option pane (Options->Options => Show tabs on the right pane) to select
it.

Rearranged panes in the consistent order (Commits, Revlist, Branches,
Repos)