Fix Windows SSH+spaces in the path issues
Adding path translation from long to short form.
This function is meaningful only on Windows: Given a long file path name,
return its short version. This is used mainly to avoid various problems
with paths containing spaces and the inability of git to handle them.

Currently the paths that are shortened are related to SSH (Paegant, PLink
Putty)