### Release 1.0.28 (2013/08/04, 13:22:01)
-------------------------------------------
SUMMARY OF CHANGES:

Implement drag and drop with the list of repos.
Now repos can be sorted manually by simply dragging and
dropping them within the list of repos. User can select
one or more repos and reorder them.

For edit repo, adding a new panel to let user be
able to edit remote repos ("Remotes")

Moved set directory in CommandPromptHere inside
exception handler since it can cause it if the
path does not exist.

Fix the LRU for the list of workspaces
File menu now shows LRU in the correct order. The top one is the
current workspace file.

Don't save workspace before clearing operation

Correcting setting 'NeedSettingsUpgrade' to be handled by
native setting provider in order to have it create necessary
folders in user app data.

Adding windows position and sizes to the list of settings
that are preserved.

Adding portable configuration.
With this change, all settings are preserved across app
update and move to different directories.

Upgrade application settings across the version increment

When adding a new remote repo, print an error message if the
repo cannot be added to git. Before this it was silently
ignoring the dialog.