### Release 1.0.31 (2015/01/27, 21:16:13)
-------------------------------------------
SUMMARY OF CHANGES:

Removed setvars build helper file. The build is done through
the Visual Studio 2012 command line that has already setup all
environment variables.

Better file handling of gitignore files
Fixing Unstage menu option
Changed shortcut for Diff vs Repo Head on an Index file from Ctrl-D to Ctrl-Z
(Ctrl-D had a race condition with the same option in the main file list)
Edit remote repo clicking on a www link recognizes github as a special case
Improved git fetch response handling on Linux

Various improvements

- Set proper monospace font for several file/text edit dialogs
- "Bitstream Vera Sans Mono" is a good font that works well on
both Windows and Linux
- Disable word-wrapping for edit boxes and enabled scroll bars
- Trim input text for remote repos UI

In the Settings/Commits, use a numeric spin box instead of a text edit
box for the number of changes to retrieve.

Trim text values in Diff, Merge and User settings input fields

In Repo Edit dialog, added a button to copy global user name and email settings; makes it simpler when creating new repos.

Set font for revision history

- Apply the same font we use for description of changes

Adding a key shortcut Ctrl+D in the pending changelists to
diff a selected file against the repo HEAD. This is the same
shortcut that does a diff in the left pane.

Removing unused "using" references in all sources

If the current repo has only one commit bundle, we don't need
to specify each file but we can simply commit all files that
are in the index.

This is also a workaround a few issues, for example, with file
rename, where deleted file needed to be amended since it was
left off.

Set a monospace font for the log window

- "Bitstream Vera Sans Mono" is a good font that works well on
both Windows and Linux