### Release 1.0.32 (2015/01/29, 21:42:22)
-------------------------------------------
SUMMARY OF CHANGES:

Fixing the issue where stopping a long git process sometimes caused the app to hung.
Changed the status font to Bitstream Vera Sans Mono; Log window font to default.

Occasionally, run the garbage collection on the loose objects in the repo.
On average, do it once after every 10 commits. This is statistical and not
guaranteed, but it is very likely that it will prevent accumulation of loose
objects in the long run and the user will not have to worry about it at all.