### Release 1.0.34 (2015/02/15, 17:11:39)
-------------------------------------------
SUMMARY OF CHANGES:

Add the ability to create repos by just dragging them onto
the repo list view. User can drag and drop one or more root
git folders and they will be added to the list.

If only a single repo is added that way, repo edit dialog will
open. More than one repo will just silently be added.

Invalid repos will be removed from the list of dragged folders.

Usability tweaks:
When starting the app, check if the global git user name and email address
are defined and if not, open the Settings dialog for the user to define them.
Also, add help tip text for that setting page.

Set AutoCloseGitOnSuccess to false by default. This is better for starters.

Adding code to check for TABs and spaces at the end of each line.
This feature is optional and by default disabled. It is controlled
in the Settings->Files along with a list of file extensions for
which it is run. By default, only these files are processed:
*.c;*.h;*.cpp;*.cc;*.cs

The Log window will display the performance of that code (in ms)
to track how long it takes over time (for devs only) and see if
the method needs performance improvements.

Git command dialog: added context menu to offer text copy
This was already present as part of the rich-text control
(Ctrl-A and Ctrl-C) but this explicit menu makes it more
obvious.

Git command dialog: Move progress indicator to the bottom and
use a wait mouse cursor while the operation is in progress.

Adding checkbox to autoclose git command dialog on the bottom
of that dialog itself since it is more convenient to have it
there.
That option is now duplicated in the settings dialog.

Internal refresh after pull or fetch; updates the Revision list.