### Release 1.0.36 (2015/09/29, 22:27:37)
-------------------------------------------
SUMMARY OF CHANGES:

Recognize a common problem where the host RSA key was not added to the list
of known hosts. Help the user by telling him that and (on Windows) offering to open the
Manage Keys dialog.

Add help buttons for remote URLs

- The pull-down is populated with all remote Push/Fetch URLs
- This is a convenience feature so the user can easily pick from the existing URLs

New Repo enable buttons to visit URL only on valid URLs

- This is for push and pop URLs, fixes targets to invalid URL pages

Save and restore Repos list columns widths

Fixed a bug when the Recreate Repos dialog was cancelled

- Ignore the retun value from this dialog since we will keep all
  modifications to the loading repos no matter how the user closes it
- Previously, the list of repos would get cleared if the user
  closed it by killing the dialog from the dialog title's 'X' button